# api-server/services/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the Nudgebee platform (Hasura actions, webhooks, integrations, observability).",
        "title": "Nudgebee Services API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/export/recommendations": {
            "post": {
                "security": [
                    {
                        "ActionToken": []
                    }
                ],
                "description": "Hasura action endpoint. Wraps an ExportRecommendationsRequest in the standard Hasura action envelope (`{action, input, session_variables}`).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "export"
                ],
                "summary": "Export recommendations as CSV/XLSX",
                "parameters": [
                    {
                        "description": "Hasura action envelope wrapping ExportRecommendationsRequest in `input.request`",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ExportRecommendationsResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ActionError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ActionError": {
            "type": "object",
            "properties": {
                "extensions": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "bad_request"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "invalid json"
                }
            }
        },
...
        "api.ExportRecommendationsResult": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "\u003cbase64-encoded payload\u003e"
                },
                "filename": {
                    "type": "string",
                    "example": "recommendations-2026-05-07.csv"
                },
                "format": {
                    "type": "string",
                    "example": "csv"
                },
                "record_count": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ActionToken": {
            "description": "Internal service-to-service auth token. Default header name (configurable via `action_api_server_token_header`); value is the `action_api_server_token` config value.",
            "type": "apiKey",
            "name": "X-ACTION-TOKEN",
            "in": "header"
        }
    }
}